graphql-schema-diff: overhaul path strings (#2397)
Until now, the paths in a change have been of the format `a.b.c`, just
nesting a path. This PR introduces a more structured representation,
along with a new string representation, that should address three
problems and allow us to represent any path within a GraphQL schema.

1. We need to be able to address items that don't have their own names.
These are schema and type extensions, as well as repeatable directives.
So I introduced indexing, like `:schema[2]`, `Query[0]`, `@tag[3]`, etc.

2. Introduce tokens to distinguish between different namespaces that
could be addressed by a path. On type definitions, we want to
distinguish paths leading to fields (or enum values, etc.) and paths to
directives on the types. To give an example, disambiguate cases like
this:
  ```graphql
  # What does the path "Test.key" refer to?
  type Test @key(fields: "id") {
    id: ID!
    key: String!
  }
  ```
  Similar problem at the top level. This is valid GraphQL:

  ```graphql
  type test { id: ID! }
  directive @test
  ```
So for directives, we have the `@` prefix. This gives us `test` vs
`@test` in the example above.

3. This results in a small grammar. So there should be a single way to
parse and print it, and that should be tested. So that is done in the
new `path` module.

This PR started as a port of
https://github.com/grafbase/api/blob/37969f88372f774f973e247174b1d5ef6c2c0f20/services/graphql/src/application/schema_proposal/comment/sdl_map/path.rs,
with a suggestion from @obmarg on a more readable syntax (`@directive`
instead of `d:directive`). I changed how indexes are represented
(`type[0]` instead of `t:type:0` for example) and had to introduce
special syntax (`&`) for interface implementations, since those were
missing.

The path module in grafbase/api will be immediately replaced with the
implementation in this PR.

closes GB-7919